+++ src/app/auth/models/user.model.ts
@@ -1,6 +1,5 @@
 /** @format */
 import { Project } from "@models/project.model";
-import { Skill } from "@office/models/skill";
 
 export class Achievement {
   id!: number;
@@ -19,8 +18,6 @@ export class User {
   avatar!: string;
   links!: string[];
   keySkills!: string[];
-  skills!: Skill[];
-  skillsIds!: number[];
   isOnline!: boolean;
   isActive!: boolean;
   member?: {
