+++ src/app/office/feed/feed.component.scss
@@ -32,4 +32,6 @@
   &__item {
     width: 100%;
   }
+
+
 }
