--- src/app/office/models/skill.ts
@@ -1,10 +1,0 @@
-/** @format */
-
-export interface Skill {
-  id: number;
-  name: string;
-  category: {
-    id: number;
-    name: string;
-  };
-}
