--- src/app/office/models/skills-group.ts
@@ -1,9 +1,0 @@
-/** @format */
-
-import { Skill } from "./skill";
-
-export interface SkillsGroup {
-  id: number;
-  name: string;
-  skills: Skill[];
-}
