--- src/app/office/models/specialization.ts
@@ -1,6 +1,0 @@
-/** @format */
-
-export interface Specialization {
-  id: number;
-  name: string;
-}
