--- src/app/office/models/specializations-group.ts
@@ -1,9 +1,0 @@
-/** @format */
-
-import { Specialization } from "./specialization";
-
-export interface SpecializationsGroup {
-  id: number;
-  name: string;
-  specializations: Specialization[];
-}
