+++ src/app/office/onboarding/onboarding/onboarding.component.scss
@@ -39,6 +39,10 @@
   &__steps {
     display: flex;
     align-items: center;
+
+    @include responsive.apply-desktop {
+      margin-left: 115px;
+    }
   }
 
   &__step {
@@ -46,20 +50,6 @@
 
     @include responsive.apply-desktop {
       margin-right: 0;
-    }
-  }
-
-  &__nav {
-    width: 100%;
-  }
-
-  &__logo {
-    display: none;
-
-    @include responsive.apply-desktop {
-      display: block;
-      width: unset;
-      margin-bottom: 30px;
     }
   }
 }
