+++ src/app/office/onboarding/stage-one/stage-one.component.scss
@@ -1,72 +1,13 @@
 @use "src/styles/responsive";
 
-.content {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 20px;
-  max-width: 1084px;
-  height: calc(100vh - 227px);
-
-  @include responsive.apply-desktop {
-    grid-template-columns: 1fr 1fr;
-    gap: 50px;
-    height: auto;
-  }
-
-  &__section {
-    display: flex;
-    flex-basis: 50%;
-    flex-direction: column;
-    min-width: 280px;
-
-    &-heading {
-      margin-bottom: 25px;
-    }
-  }
-
-  &__right {
-    grid: unset;
-    height: calc(80vh - 296px);
-    padding-bottom: 14px;
-    background-color: var(--white);
-    border: 1px solid var(--medium-grey-for-outline);
-    border-radius: 8px;
-    box-shadow: 5px 5px 25px 0 var(--gray-for-shadow);
-
+.page {
+  &__form {
     @include responsive.apply-desktop {
-      grid-row: 1/3;
-      grid-column: 2/3;
-      height: 480px;
-    }
-
-    &-heading {
-      padding: 14px 14px 0;
+      width: 415px;
     }
   }
 
-  &__actions {
-    grid-column: 1/2;
-    order: 3;
-    margin-top: auto;
-  }
-
-  &__specs-groups {
-    height: 100%;
-    overflow: auto;
-    scrollbar-width: thin;
-
-    &-list {
-      display: flex;
-      flex-direction: column;
-      gap: 20px;
-      padding: 0 14px;
-    }
-
-    &-list-item {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      cursor: pointer;
-    }
+  &__field {
+    margin-bottom: 12px;
   }
 }
