+++ src/app/office/onboarding/stage-one/stage-one.component.ts
@@ -1,22 +1,16 @@
 /** @format */
 
-import { Component, OnDestroy, OnInit, signal } from "@angular/core";
-import { NonNullableFormBuilder, ReactiveFormsModule, Validators } from "@angular/forms";
-import { map, Observable, Subscription, take } from "rxjs";
+import { Component, OnDestroy, OnInit } from "@angular/core";
+import { FormArray, FormBuilder, FormGroup, ReactiveFormsModule, Validators } from "@angular/forms";
+import { ErrorMessage } from "@error/models/error-message";
+import { concatMap, Subscription, take } from "rxjs";
 import { AuthService } from "@auth/services";
 import { ValidationService } from "@core/services";
-import { ActivatedRoute, Router } from "@angular/router";
+import { Router } from "@angular/router";
 import { OnboardingService } from "../services/onboarding.service";
 import { ControlErrorPipe } from "@core/pipes/control-error.pipe";
 import { ButtonComponent, IconComponent, InputComponent } from "@ui/components";
 import { TagComponent } from "@ui/components/tag/tag.component";
-import { CommonModule } from "@angular/common";
-import { AutoCompleteInputComponent } from "@ui/components/autocomplete-input/autocomplete-input.component";
-import { SpecializationsGroup } from "@office/models/specializations-group";
-import { SpecializationsGroupComponent } from "@office/shared/specializations-group/specializations-group.component";
-import { Specialization } from "@office/models/specialization";
-import { SpecializationsService } from "@office/services/specializations.service";
-import { ErrorMessage } from "@error/models/error-message";
 
 @Component({
   selector: "app-stage-one",
@@ -30,81 +24,85 @@ import { ErrorMessage } from "@error/models/error-message";
     IconComponent,
     ButtonComponent,
     ControlErrorPipe,
-    InputComponent,
-    AutoCompleteInputComponent,
-    SpecializationsGroupComponent,
-    CommonModule,
   ],
 })
 export class OnboardingStageOneComponent implements OnInit, OnDestroy {
   constructor(
-    private readonly nnFb: NonNullableFormBuilder,
+    private readonly fb: FormBuilder,
     private readonly authService: AuthService,
     private readonly onboardingService: OnboardingService,
     private readonly validationService: ValidationService,
-    private readonly specsService: SpecializationsService,
-    private readonly router: Router,
-    private readonly route: ActivatedRoute
-  ) {}
-
-  stageForm = this.nnFb.group({
-    speciality: ["", Validators.required],
-  });
-
-  nestedSpecializations$: Observable<SpecializationsGroup[]> = this.route.data.pipe(
-    map(r => r["data"])
-  );
-
-  inlineSpecializations = signal<Specialization[]>([]);
-
-  stageSubmitting = signal(false);
-
-  errorMessage = ErrorMessage;
-
-  subscriptions$ = signal<Subscription[]>([]);
+    private readonly router: Router
+  ) {
+    this.stageForm = this.fb.group({
+      speciality: [""],
+      keySkills: this.fb.array([]),
+    });
+  }
 
   ngOnInit(): void {
     const formValueState$ = this.onboardingService.formValue$.pipe(take(1)).subscribe(fv => {
       this.stageForm.patchValue({
         speciality: fv.speciality,
+        keySkills: fv.keySkills,
       });
+
+      fv.keySkills?.forEach(skill => this.addKeySkill(skill));
     });
 
     const formValueChange$ = this.stageForm.valueChanges.subscribe(value => {
       this.onboardingService.setFormValue(value);
     });
 
-    this.subscriptions$().push(formValueState$, formValueChange$);
+    this.subscriptions$.push(formValueState$, formValueChange$);
   }
 
   ngOnDestroy(): void {
-    this.subscriptions$().forEach($ => $.unsubscribe());
+    this.subscriptions$.forEach($ => $.unsubscribe());
   }
+
+  subscriptions$: Subscription[] = [];
+  stageForm: FormGroup;
+
+  newKeySkillTitle = "";
+
+  get keySkills(): FormArray {
+    return this.stageForm.get("keySkills") as FormArray;
+  }
+
+  addKeySkill(title?: string): void {
+    const fromState = title ?? this.newKeySkillTitle;
+    if (!fromState) {
+      return;
+    }
+
+    const control = this.fb.control(fromState, [Validators.required]);
+    this.keySkills.push(control);
+
+    this.newKeySkillTitle = "";
+  }
+
+  removeKeySkill(i: number): void {
+    this.keySkills.removeAt(i);
+  }
+
+  errorMessage = ErrorMessage;
+  stageSubmitting = false;
 
   onSubmit(): void {
     if (!this.validationService.getFormValidation(this.stageForm)) {
       return;
     }
 
-    this.stageSubmitting.set(true);
+    this.stageSubmitting = true;
 
-    this.authService.saveProfile(this.stageForm.value).subscribe(() => {
-      this.router
-        .navigateByUrl("/office/onboarding/stage-2")
-        .then(() => console.debug("Route changed from OnboardingStageOneComponent"));
-    });
-  }
-
-  onSelectSpec(speciality: Specialization): void {
-    this.stageForm.patchValue({ speciality: speciality.name });
-  }
-
-  onSearchSpec(query: string): void {
-    this.specsService
-      .getSpecializationsInline(query, 1000, 0)
-      .pipe(take(1))
-      .subscribe(({ results }) => {
-        this.inlineSpecializations.set(results);
+    this.authService
+      .saveProfile(this.stageForm.value)
+      .pipe(concatMap(() => this.authService.setOnboardingStage(2)))
+      .subscribe(() => {
+        this.router
+          .navigateByUrl("/office/onboarding/stage-2")
+          .then(() => console.debug("Route changed from OnboardingStageOneComponent"));
       });
   }
 }
