--- src/app/office/onboarding/stage-three/stage-three.component.html
@@ -1,50 +1,0 @@
-<!-- @format -->
-
-<div class="page">
-  <div class="auth__greeting">
-    <h3 class="auth__title">Выбери, кем ты будешь на платформе</h3>
-    <p class="auth__info">Выбери, кто ты?</p>
-  </div>
-  <div class="page__list">
-    <app-user-type-card [isActive]="userRole === 1" (click)="onSetRole(1)">
-      <ng-container title>Участник</ng-container>
-      <ng-container description>
-        Участник на платформе - это школьник или студент, который может зарегистрировать свой проект
-        или вступить в уже существующий, получить поддержку от ментора, участвовать в мероприятиях,
-        читать новости
-      </ng-container>
-    </app-user-type-card>
-    <app-user-type-card [isActive]="userRole === 2" (click)="onSetRole(2)">
-      <ng-container title>Инвестор</ng-container>
-      <ng-container description>
-        На платформе инвестор - человек, который владеет финансами и может как профинансировать
-        проект, так и создать свой проект на платформе. К инвестору могут обратиться участники и
-        проекты.
-      </ng-container>
-    </app-user-type-card>
-    <app-user-type-card [isActive]="userRole === 3" (click)="onSetRole(3)">
-      <ng-container title>Ментор</ng-container>
-      <ng-container description>
-        На платформе ментор - человек, который открыт дать помощь инновационной молодежи, к ментору
-        могут обратиться как участники, так и проекты. Сам ментор может написать в проект и
-        поддержать его.
-      </ng-container>
-    </app-user-type-card>
-    <app-user-type-card [isActive]="userRole === 4" (click)="onSetRole(4)">
-      <ng-container title>Эксперт</ng-container>
-      <ng-container description>
-        На платформе эксперт - человек, который открыт дать экспертизу из какой-либо сферы. К
-        эксперту могут обратиться как участник, так и проекты. Сам эксперт может написать в проект и
-        поддержать его.
-      </ng-container>
-    </app-user-type-card>
-  </div>
-  <app-button class="page__button" [loader]="stageSubmitting" (click)="onSubmit()"
-    >Продолжить</app-button
-  >
-  @if (stageTouched && userRole === -1) {
-  <p class="error page__error">
-    Вы ничего не выбрали, выберите подходящий пункт и нажмите “продолжить”
-  </p>
-  }
-</div>
