--- src/app/office/onboarding/stage-three/stage-three.component.scss
@@ -1,40 +1,0 @@
-@use "src/styles/responsive";
-@use "src/styles/typography";
-
-.page {
-  &__list {
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-gap: 12px;
-    height: 318px;
-    padding-bottom: 20px;
-    margin-top: 20px;
-    overflow-y: auto;
-
-    @include responsive.apply-desktop {
-      grid-template-columns: 1fr 1fr;
-      grid-gap: 20px;
-      max-width: 820px;
-      height: unset;
-      padding-bottom: 0;
-      margin-bottom: 31px;
-    }
-  }
-
-  &__button {
-    display: block;
-
-    @include responsive.apply-desktop {
-      width: 415px;
-    }
-  }
-
-  &__error {
-    margin-top: 16px;
-    text-align: center;
-
-    @include responsive.apply-desktop {
-      max-width: 415px;
-    }
-  }
-}
