+++ src/app/office/onboarding/stage-two/stage-two.component.scss
@@ -1,99 +1,40 @@
 @use "src/styles/responsive";
+@use "src/styles/typography";
 
-.content {
-  display: grid;
-  grid-template-rows: 1fr;
-  grid-template-columns: 1fr;
-  gap: 20px;
-  max-width: 1084px;
-  height: calc(100vh - 227px);
-  overflow: hidden;
-
-  @include responsive.apply-desktop {
-    grid-template-columns: 1fr 1fr;
-    row-gap: 20px;
-    column-gap: 50px;
-    max-height: 480px;
-  }
-
-  &__section {
-    display: flex;
-    flex-basis: 50%;
-    flex-direction: column;
-    min-width: 280px;
-    height: 100%;
-    overflow: hidden;
-
-    &-heading {
-      margin-bottom: 25px;
-    }
-  }
-
-  &__left,
-  &__search,
-  &__skills {
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__left {
-    gap: 20px;
-  }
-
-  &__skills {
-    height: 100%;
-    overflow: hidden;
-
-    &-basket {
-      overflow: hidden;
-    }
-  }
-
-  &__right {
-    display: none !important;
-    grid: unset;
-    height: calc(80vh - 296px);
-    padding-bottom: 14px;
-    background-color: var(--white);
-    border: 1px solid var(--medium-grey-for-outline);
-    border-radius: 8px;
-    box-shadow: 5px 5px 25px 0 var(--gray-for-shadow);
+.page {
+  &__list {
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-gap: 12px;
+    height: 318px;
+    padding-bottom: 20px;
+    margin-top: 20px;
+    overflow-y: auto;
 
     @include responsive.apply-desktop {
-      display: flex !important;
-      grid-row: 1/3;
-      grid-column: 2/3;
-      height: 100%;
-    }
-
-    &-heading {
-      padding: 14px 14px 0;
+      grid-template-columns: 1fr 1fr;
+      grid-gap: 20px;
+      max-width: 820px;
+      height: unset;
+      padding-bottom: 0;
+      margin-bottom: 31px;
     }
   }
 
-  &__actions {
-    grid-column: 1/2;
-    order: 3;
-    margin-top: auto;
+  &__button {
+    display: block;
+
+    @include responsive.apply-desktop {
+      width: 415px;
+    }
   }
 
-  &__skills-groups {
-    height: 100%;
-    overflow: auto;
-    scrollbar-width: thin;
+  &__error {
+    margin-top: 16px;
+    text-align: center;
 
-    &-list {
-      display: flex;
-      flex-direction: column;
-      gap: 20px;
-      padding: 0 14px;
-    }
-
-    &-list-item {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      cursor: pointer;
+    @include responsive.apply-desktop {
+      max-width: 415px;
     }
   }
 }
