+++ src/app/office/onboarding/stage-two/stage-two.component.spec.ts
@@ -4,30 +4,27 @@ import { ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { OnboardingStageTwoComponent } from "./stage-two.component";
 import { of } from "rxjs";
-import { ReactiveFormsModule } from "@angular/forms";
 import { AuthService } from "@auth/services";
+import { OnboardingService } from "../services/onboarding.service";
 import { RouterTestingModule } from "@angular/router/testing";
-import { HttpClientTestingModule } from "@angular/common/http/testing";
 
 describe("StageTwoComponent", () => {
   let component: OnboardingStageTwoComponent;
   let fixture: ComponentFixture<OnboardingStageTwoComponent>;
 
   beforeEach(async () => {
-    const authSpy = {
-      profile: of({}),
+    const authSpy = jasmine.createSpyObj({
       saveProfile: of({}),
       setOnboardingStage: of({}),
-    };
+    });
+    const onboardingSpy = jasmine.createSpyObj({}, { formValue$: of({}) });
 
     await TestBed.configureTestingModule({
-      imports: [
-        ReactiveFormsModule,
-        RouterTestingModule,
-        HttpClientTestingModule,
-        OnboardingStageTwoComponent,
+      imports: [RouterTestingModule, OnboardingStageTwoComponent],
+      providers: [
+        { provide: AuthService, useValue: authSpy },
+        { provide: OnboardingService, useValue: onboardingSpy },
       ],
-      providers: [{ provide: AuthService, useValue: authSpy }],
     }).compileComponents();
   });
