+++ src/app/office/onboarding/stage-two/stage-two.component.ts
@@ -1,127 +1,64 @@
 /** @format */
 
-import { Component, OnDestroy, OnInit, signal } from "@angular/core";
-import { NonNullableFormBuilder, ReactiveFormsModule, Validators } from "@angular/forms";
-import { concatMap, map, Observable, Subscription, take } from "rxjs";
+import { Component, OnDestroy, OnInit } from "@angular/core";
 import { AuthService } from "@auth/services";
-import { ValidationService } from "@core/services";
-import { ActivatedRoute, Router } from "@angular/router";
+import { concatMap, Subscription, take } from "rxjs";
+import { Router } from "@angular/router";
 import { OnboardingService } from "../services/onboarding.service";
-import { ControlErrorPipe } from "@core/pipes/control-error.pipe";
-import { ButtonComponent, IconComponent } from "@ui/components";
-import { CommonModule } from "@angular/common";
-import { AutoCompleteInputComponent } from "@ui/components/autocomplete-input/autocomplete-input.component";
-import { Skill } from "@office/models/skill";
-import { SkillsService } from "@office/services/skills.service";
-import { SkillsGroup } from "@office/models/skills-group";
-import { SkillsGroupComponent } from "@office/shared/skills-group/skills-group.component";
-import { SkillsBasketComponent } from "@office/shared/skills-basket/skills-basket.component";
+import { ButtonComponent } from "@ui/components";
+import { UserTypeCardComponent } from "../user-type-card/user-type-card.component";
 
 @Component({
   selector: "app-stage-two",
   templateUrl: "./stage-two.component.html",
   styleUrl: "./stage-two.component.scss",
   standalone: true,
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    IconComponent,
-    ButtonComponent,
-    ControlErrorPipe,
-    AutoCompleteInputComponent,
-    SkillsGroupComponent,
-    SkillsBasketComponent,
-  ],
+  imports: [UserTypeCardComponent, ButtonComponent],
 })
 export class OnboardingStageTwoComponent implements OnInit, OnDestroy {
   constructor(
-    private readonly nnFb: NonNullableFormBuilder,
     private readonly authService: AuthService,
     private readonly onboardingService: OnboardingService,
-    private readonly validationService: ValidationService,
-    private readonly skillsService: SkillsService,
-    private readonly router: Router,
-    private readonly route: ActivatedRoute
+    private readonly router: Router
   ) {}
 
-  stageForm = this.nnFb.group({
-    skills: this.nnFb.control<Skill[]>([], Validators.required),
-  });
-
-  nestedSkills$: Observable<SkillsGroup[]> = this.route.data.pipe(map(r => r["data"]));
-
-  searchedSkills = signal<Skill[]>([]);
-
-  stageSubmitting = signal(false);
-
-  subscriptions$ = signal<Subscription[]>([]);
-
   ngOnInit(): void {
-    const fv$ = this.onboardingService.formValue$
-      .pipe(take(1))
-      .subscribe(({ skills }) => this.stageForm.patchValue({ skills }));
-
-    const formValueChange$ = this.stageForm.valueChanges.subscribe(value => {
-      this.onboardingService.setFormValue(value);
+    const formValue$ = this.onboardingService.formValue$.pipe(take(1)).subscribe(fv => {
+      this.userRole = fv.userType ? fv.userType : -1;
     });
 
-    this.subscriptions$().push(fv$, formValueChange$);
+    this.subscriptions$.push(formValue$);
   }
 
   ngOnDestroy(): void {
-    this.subscriptions$().forEach($ => $.unsubscribe());
+    this.subscriptions$.forEach($ => $.unsubscribe());
   }
 
-  onSubmit(): void {
-    if (!this.validationService.getFormValidation(this.stageForm)) {
+  userRole!: number;
+  stageTouched = false;
+  stageSubmitting = false;
+  subscriptions$: Subscription[] = [];
+
+  onSetRole(role: number) {
+    this.userRole = role;
+    this.onboardingService.setFormValue({ userType: role });
+  }
+
+  onSubmit() {
+    if (this.userRole === -1) {
+      this.stageTouched = true;
       return;
     }
 
-    this.stageSubmitting.set(true);
-
-    const { skills } = this.stageForm.getRawValue();
+    this.stageSubmitting = true;
 
     this.authService
-      .saveProfile({ skillsIds: skills.map(skill => skill.id) })
-      .pipe(concatMap(() => this.authService.setOnboardingStage(2)))
+      .saveProfile({ userType: this.userRole })
+      .pipe(concatMap(() => this.authService.setOnboardingStage(null)))
       .subscribe(() => {
         this.router
-          .navigateByUrl("/office/onboarding/stage-3")
-          .then(() => console.debug("Route changed from OnboardingStageOneComponent"));
+          .navigateByUrl("/office")
+          .then(() => console.debug("Route changed from OnboardingStageTwo"));
       });
-  }
-
-  onAddSkill(newSkill: Skill): void {
-    const { skills } = this.stageForm.getRawValue();
-
-    const isPresent = skills.some(s => s.id === newSkill.id);
-
-    if (isPresent) return;
-
-    this.stageForm.patchValue({ skills: [...skills, newSkill] });
-  }
-
-  onRemoveSkill(oddSkill: Skill): void {
-    const { skills } = this.stageForm.getRawValue();
-
-    this.stageForm.patchValue({ skills: skills.filter(skill => skill.id !== oddSkill.id) });
-  }
-
-  onOptionToggled(toggledSkill: Skill): void {
-    const { skills } = this.stageForm.getRawValue();
-
-    const isPresent = skills.some(skill => skill.id === toggledSkill.id);
-
-    if (isPresent) {
-      this.onRemoveSkill(toggledSkill);
-    } else {
-      this.onAddSkill(toggledSkill);
-    }
-  }
-
-  onSearchSkill(query: string): void {
-    this.skillsService
-      .getSkillsInline(query, 1000, 0)
-      .subscribe(({ results }) => this.searchedSkills.set(results));
   }
 }
