--- src/app/office/onboarding/stage-two/stage-two.resolver.ts
@@ -1,12 +1,0 @@
-/** @format */
-
-import { inject } from "@angular/core";
-import type { ResolveFn } from "@angular/router";
-import { SkillsService } from "@office/services/skills.service";
-import { SkillsGroup } from "@office/models/skills-group";
-
-export const StageTwoResolver: ResolveFn<SkillsGroup[]> = () => {
-  const skillsService = inject(SkillsService);
-
-  return skillsService.getSkillsNested();
-};
