--- src/app/office/shared/skills-basket/skills-basket.component.html
@@ -1,23 +1,0 @@
-<!-- @format -->
-
-<div class="basket" [class.basket--error]="error">
-  <div class="basket__content" [class.basket__content--empty]="!value().length">
-    @for (skill of value(); track skill.id) {
-    <div class="basket__skill">
-      <span class="text-body-14">{{ skill.name }}</span>
-      <i
-        appIcon
-        class="basket__skill-delete-icon"
-        icon="basket"
-        appSquare="18"
-        (click)="deleteSkill(skill.id)"
-      ></i>
-    </div>
-    } @empty {
-    <div class="basket__empty-placeholder">
-      <img src="/assets/images/office/onboarding/laptop-guy.svg" alt="" />
-      <span class="text-body-14">Выберите навыки</span>
-    </div>
-    }
-  </div>
-</div>
