--- src/app/office/shared/skills-basket/skills-basket.component.scss
@@ -1,57 +1,0 @@
-@use "src/styles/responsive";
-@use "src/styles/typography";
-
-:host {
-  height: 100%;
-}
-
-.basket {
-  height: 100%;
-  padding: 12px 0 12px 12px;
-  overflow: hidden;
-  border: 1px solid var(--medium-grey-for-outline);
-  border-radius: 15px;
-
-  &__content {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 10px;
-    align-content: baseline;
-    height: 100%;
-    padding-right: 12px;
-    overflow-y: auto;
-    scrollbar-width: thin;
-
-    &--empty {
-      align-content: unset;
-    }
-  }
-
-  &__skill {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 5px 10px;
-    color: var(--white);
-    background-color: var(--accent);
-    border-radius: var(--rounded-md);
-
-    &-delete-icon {
-      margin-left: 10px;
-      cursor: pointer;
-    }
-  }
-
-  &__empty-placeholder {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    align-items: center;
-    align-self: center;
-    margin: 0 auto;
-  }
-
-  &--error {
-    border-color: var(--red);
-  }
-}
