--- src/app/office/shared/skills-group/skills-group.component.html
@@ -1,19 +1,0 @@
-<!-- @format -->
-<div class="heading" [class.heading--open]="contentVisible" (click)="toggleContentVisible()">
-  <span class="text-body-16">{{ title }}</span>
-  <i appIcon icon="arrow-no-body" appWidth="18" appHeight="9" appViewBox="0 0 18 9"></i>
-</div>
-@if (contentVisible()) {
-<div class="content">
-  @for (opt of options; track opt.id) {
-  <div class="content__option" (click)="optionToggled.emit(opt)">
-    <div class="checkbox">
-      <div class="checkbox__field" [class.checkbox__field--checked]="opt.checked">
-        <i appIcon icon="check" appWidth="12" appHeight="10"></i>
-      </div>
-    </div>
-    <label class="text-body-16 content__option-label">{{ opt.name }}</label>
-  </div>
-  }
-</div>
-}
