--- src/app/office/shared/skills-group/skills-group.component.scss
@@ -1,65 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  width: 100%;
-  min-height: 32px;
-
-  .heading {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    height: 32px;
-    cursor: pointer;
-
-    i {
-      transform: rotate(90deg);
-    }
-
-    &--open {
-      i {
-        transform: rotate(180deg);
-      }
-    }
-  }
-
-  .content {
-    display: flex;
-    flex-direction: column;
-    gap: 14px;
-
-    &__option {
-      display: flex;
-      gap: 10px;
-      align-items: center;
-      cursor: pointer;
-
-      &-label {
-        cursor: pointer;
-      }
-    }
-  }
-
-  .checkbox {
-    &__field {
-      position: relative;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 20px;
-      height: 20px;
-      border: 2px solid var(--gray);
-      border-radius: var(--rounded-sm);
-
-      &--checked {
-        background-color: var(--accent);
-        border-color: var(--accent);
-      }
-    }
-
-    i {
-      color: var(--white);
-    }
-  }
-}
