--- src/app/office/shared/specializations-group/specializations-group.component.html
@@ -1,18 +1,0 @@
-<!-- @format -->
-<div
-  class="heading"
-  [class.heading--open]="contentVisible"
-  (click)="contentVisible = !contentVisible"
->
-  <span class="text-body-16">{{ title }}</span>
-  <i appIcon icon="arrow-no-body" appWidth="18" appHeight="9" appViewBox="0 0 18 9"></i>
-</div>
-@if (contentVisible) {
-<div class="content">
-  @for (opt of options; track opt.id) {
-  <div class="content__option" (click)="onSelectOption(opt)">
-    <span class="text-body-14">{{ opt.name }}</span>
-  </div>
-  }
-</div>
-}
