--- src/app/ui/components/autocomplete-input/autocomplete-input.component.scss
@@ -1,127 +1,0 @@
-@use "src/styles/typography";
-
-.field {
-  position: relative;
-
-  &__input {
-    position: relative;
-    z-index: 4;
-    width: 100%;
-    padding: 13px 80px 13px 20px;
-    color: var(--black);
-    background-color: var(--white);
-    border: 1px solid var(--gray);
-    border-radius: var(--rounded-lg);
-    outline: none;
-    transition: all 0.2s;
-
-    &::placeholder {
-      color: var(--dark-grey);
-    }
-
-    &--error {
-      border-color: var(--red);
-    }
-
-    &:focus {
-      border-color: var(--accent);
-
-      +.field__arrow {
-        color: var(--accent);
-
-        i {
-          transform: rotate(0);
-        }
-      }
-    }
-
-    @include typography.body-16;
-  }
-
-  &__dropdown {
-    position: absolute;
-    top: 80%;
-    right: 0;
-    left: 0;
-    z-index: 3;
-    padding-top: 10px;
-    background-color: transparent;
-    box-shadow: 5px 5px 25px 0 var(--gray-for-shadow);
-    transform-origin: center top;
-
-    &--options,
-    &--no-results {
-      max-height: 300px;
-      padding: 10px;
-      overflow-y: auto;
-      background-color: var(--white);
-    }
-  }
-
-  &__option {
-    padding: 7px;
-    color: var(--dark-grey);
-    cursor: pointer;
-    border-radius: var(--rounded-lg);
-    transition: background-color 0.2s;
-
-    &:hover {
-      background-color: var(--light-gray);
-    }
-
-    @include typography.body-14;
-  }
-
-  &__chip {
-    position: absolute;
-    top: 50%;
-    left: 10px;
-    z-index: 4;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 5px 10px;
-    color: var(--white);
-    background-color: var(--accent);
-    border-radius: var(--rounded-md);
-    transform: translateY(-50%);
-
-    i {
-      margin-left: 10px;
-      cursor: pointer;
-    }
-  }
-
-  &__search-icon {
-    position: absolute;
-    top: 50%;
-    right: 10px;
-    z-index: 4;
-    transform: translateY(-50%);
-  }
-
-  &__loading {
-    position: absolute;
-    top: 50%;
-    right: 45px;
-    z-index: 4;
-    transform: translateY(-50%);
-  }
-}
-
-.blink {
-  animation: blink 2s infinite both;
-}
-
-@keyframes blink {
-  0%,
-  50%,
-  100% {
-    opacity: 1;
-  }
-
-  25%,
-  75% {
-    opacity: 0;
-  }
-}
