+++ src/styles/_colors.scss
@@ -17,7 +17,6 @@
   --dark-grey: #8c888a;
   --gray: #d3d3d3;
   --light-gray: #f9f9f9;
-  --gray-for-shadow: rgb(159 159 159 / 15%);
   --grey-button: #e5e5e5e5;
   --medium-grey-for-outline: #eee;
   --grey-for-text: #a59fb9;
